17_sites_monitoring/requirements.txt
pytz==2018.7
requests==2.20.1
python-whois==0.7.0